fix: resolve clippy warnings across gpu, runtime, and main crates
- gpu: add crate-level allow for missing_safety_doc on FFI functions
- gpu: mark plugin_init/plugin_describe as unsafe extern "C"
- gpu: remove unnecessary pointer casts in buffer.rs and compile.rs
- gpu: replace format!("literal") with "literal".to_string() in msl.rs
- gpu: add Default impl and is_empty() for KernelCache, use entry API
- gpu: use div_ceil() instead of manual ceiling division in ops.rs
- gpu: fix needless_range_loop and manual_slice_size_calculation in tests
- runtime: fix needless_range_loop in tensor shape copy
- main: use iter().flatten() instead of manual Option unwrapping

+++ gpu/src/metal/compile.rs
@@ -44,7 +44,7 @@ pub fn compile_msl(
         .newComputePipelineStateWithFunction_error(&function)
         .map_err(|e| format!("pipeline creation failed: {}", e))?;
 
-    let max_threads_per_group = pipeline.maxTotalThreadsPerThreadgroup() as usize;
+    let max_threads_per_group = pipeline.maxTotalThreadsPerThreadgroup();
 
     Ok(CompiledKernel {
         pipeline,
